auth ^2
- checked authentication on routes
- added authorisation on relevant routes
- added authorisation test to /users/all to only send a message, as a test
- required further testing.

+++ src/controllers/UserController.js
@@ -37,15 +37,31 @@ const errorSwtich = (err, response) => {
 /* === USER GET ROUTES === */
 
 // GET all users in db
+// Has authentication and authorisation (just to test - all authenticated users can access)
 // eg GET localhost:3000/users/all/
 router.get('/all', auth, async (request, response) => {
-  
+  // extract id from request, to return name, as test
+  const _id = request.userId;
+
   try {
-    const result = await User.find({});
+    const result = await User.findById(_id);
+    const resultAll = await User.find();
     
-    response.json({
-      Users: result
-    });
+    if (!result) {
+      return response.status(400).send('User does not exist.');
+    }
+
+    if (result.isAdmin) {
+      response.json({
+        message: "Hello Admin",
+        Users: resultAll
+       })
+    } else {
+      response.json({
+        message: "Hello non-Admin",
+        Users: resultAll
+      });
+    }
     
   } catch (err) {
     errorSwtich(err, response);
@@ -56,11 +72,14 @@ router.get('/all', auth, async (request, response) => {
 // GET all FULL users in db
 // eg GET localhost:3000/users/all/fullusers
 router.get('/all/fullusers', auth, async (request, response) => {
+  const _id = request.userId;
+
   try {
-    const result = await User.find({isFullUser: true});
+    const result = await User.findById(_id);
+    const outcome = await User.find({isFullUser: true});
     
     response.json({
-      Users: result
+      Users: outcome
     });
     
   } catch (err) {
@@ -72,11 +91,14 @@ router.get('/all/fullusers', auth, async (request, response) => {
 // GET user by id
 // eg: GET localhost:3000/users/one/id/5e9b2b7b9b9b9b9b9b9b9b9b
 router.get('/one/id/:id', auth, async (request, response) => {
+  const _id = request.userId;
+
   try {
-    const result = await User.findById(request.params.id);
+    const result = await User.findById(_id);
+    const outcome = await User.findById(request.params.id);
     
     response.json({
-      User: result
+      User: outcome
     });
     
   } catch (err) {
@@ -87,16 +109,19 @@ router.get('/one/id/:id', auth, async (request, response) => {
 // GET user by name
 // eg: GET localhost:3000/users/one/name/katie/lock
 router.get('/one/name/:firstName/:lastName', auth, async (request, response) => {
-  try {
-    // Removed case sensitivity from query params
+  const _id = request.userId;
 
+  try {
+    const result = await User.findById(_id);
+    
+    // Removes case sensitivity from query params
     const firstNameRegex = new RegExp(request.params.firstName, 'i'); // 'i' makes it case insensitive
     const lastNameRegex = new RegExp(request.params.lastName, 'i'); // 'i' makes it case insensitive
     
-    const result = await User.findOne({ 'name.first': { $regex: firstNameRegex }, 'name.last': { $regex: lastNameRegex } });
+    const outcome = await User.findOne({ 'name.first': { $regex: firstNameRegex }, 'name.last': { $regex: lastNameRegex } });
     
     response.json({
-      Users: result
+      Users: outcome
     });
     
   } catch (err) {
@@ -107,9 +132,12 @@ router.get('/one/name/:firstName/:lastName', auth, async (request, response) => 
 // GET user regex, by name - not case sensitive
 // eg: GET localhost:3000/users/search/katie
 router.get('/search/:string', auth, async (request, response) => {
+  const _id = request.userId;
+
   try {
+    const result = await User.findById(_id);
     const regex = new RegExp(request.params.string, 'i');
-    const result = await User.find({ 
+    const outcome = await User.find({ 
       $or: [
         { 'name.first': { $regex: regex } },
         { 'name.last': { $regex: regex } }
@@ -117,7 +145,7 @@ router.get('/search/:string', auth, async (request, response) => {
     });
     
     response.json({
-      Users: result
+      Users: outcome
     });
     
   } catch (err) {
@@ -127,12 +155,21 @@ router.get('/search/:string', auth, async (request, response) => {
 
 
 // GET user by Manager id
+// Has authentication and authorisation = one of isLineManager, isSeniorManager, isAdmin must be true
+// eg: GET localhost:3000/users/all/manager/657d7c33d8b97e77efe01939
 router.get('/all/manager/:id', auth, async (request, response) => {
+  const _id = request.userId;
+
   try {
-    const result = await User.find({ lineManagerId: request.params.id} );
+    const result = await User.findById(_id);
+    const outcome = await User.find({ lineManagerId: request.params.id} );
+    
+    if (!result.isLineManager && !result.isSeniorManager && !result.isAdmin) {
+      return response.status(400).send('You are not authorised to see this. Close your eyes and walk away.');
+    }
     
     response.json({
-      User: result
+      User: outcome
     });
     
   } catch (err) {
@@ -144,17 +181,18 @@ router.get('/all/manager/:id', auth, async (request, response) => {
 // eg: GET localhost:3000/users/one/email/jordan.benjamin@yourcompany.com
 
 router.get('/one/email/:email', auth, async (request, response) => {
+  const _id = request.userId;
 
   try {
-    // Removed case sensitivity from query params
-    const { email } = request.body;
-
-    const emailRegex = new RegExp(email, 'i'); // 'i' makes it case insensitive
+    const result = await User.findById(_id);
+    const { email } = request.params;
     
-    const result = await User.findOne({ 'email': { $regex: emailRegex } });
+    // Removes case sensitivity from query params
+    const emailRegex = new RegExp(email, 'i'); // 'i' makes it case insensitive
+    const outcome = await User.findOne({ 'email': { $regex: emailRegex } });
     
     response.json({
-      Users: result
+      Users: outcome
     });
     
   } catch (err) {
@@ -167,15 +205,23 @@ router.get('/one/email/:email', auth, async (request, response) => {
 
 
 // POST new user
+// Has authentication and authorisation = isAdmin must be true
 // eg: POST http://localhost:3000/users/new
-/* ADD AUTHORISATION */
+
 router.post('/new', auth, async (request, response) => {
+  const _id = request.userId;
+
   try {
+    const result = await User.findById(_id);
+    if(!result.isAdmin) {
+      return response.status(403).send('You are not authorised to do that. Wash your mouth with soap.');
+    };
+
     const newUser = new User(request.body);
-    const result = await newUser.save();
+    const outcome = await newUser.save();
     
     response.json({
-      User: result
+      User: outcome
     });
     
   } catch (err) {
@@ -188,9 +234,12 @@ router.post('/new', auth, async (request, response) => {
 
 // PATCH user by id, for use by self
 // eg: PATCH localhost:3000/users/update/self/5e9b2b7b9b9b9b9b9b9b9b9b
-/* ADD AUTHORISATION */
+/* NEEDS TESTING */
 router.patch('/update/self/:id', auth, async (request, response) => {
+  const _id = request.userId;
+
   try {
+    const result = await User.findById(_id);
     // creates object of keys from request body
     const updates = Object.keys(request.body);
     // so as to limit update to only allowed updates: userTagLine
@@ -198,8 +247,8 @@ router.patch('/update/self/:id', auth, async (request, response) => {
     // applies updated values to user object
     const isValid = updates.every((update) => allowedUpdates.includes(update));
 
-    // if trying to update a non-updatable field, return error
-    if (!isValid) {
+    // if trying to update a non-updatable field, trying to update a profile other than their own, or is not admin return error
+  if (!isValid || request.params.id !== _id || !result.isAdmin) {
       return response.status(400).send({ error: 'Your intent is good but we can\'t update that.' });
     }
 
@@ -224,10 +273,13 @@ router.patch('/update/self/:id', auth, async (request, response) => {
 
 
 // PATCH admin by id 
+// Has authentication and authorisation = isAdmin must be true
 // eg: PATCH localhost:3000/users/update/admin/5e9b2b7b9b9b9b9b9b9b9b9b
-/* ADD AUTHORISATION */
 router.patch('/update/admin/:id', auth, async (request, response) => {
+  const _id = request.userId;
+  
   try {
+    const result = await User.findById(_id);
     // creates object of keys from request body
     const updates = Object.keys(request.body);
     // so as to limit update to only allowed updates: userTagLine
@@ -237,7 +289,7 @@ router.patch('/update/admin/:id', auth, async (request, response) => {
     const isValid = updates.every((update) => allowedUpdates.includes(update));
 
     // if trying to update a non-updatable field, return error
-    if (!isValid) {
+    if (!request.isAdmin || !isValid) {
       return response.status(400).send({ error: 'Your intent is good but there is something in there that we can\'t update.' });
     }
 
@@ -267,18 +319,24 @@ router.patch('/update/admin/:id', auth, async (request, response) => {
 // eg: DELETE localhost:3000/admin/delete/5e9b2b7b9b9b9b9b9b9b9b9b
 /* ADD AUTHORISATION */
 router.delete('/delete/admin/:id', auth, async (request, response) => {
+  const _id = request.userId;
+
   try {
-    const result = await User.findByIdAndDelete(request.params.id);
+    const result = await User.findById(_id);
+    if (!result.isAdmin) {
+      return response.status(403).send('You are not authorised to do that. Wash your mouth with soap.');
+    }
+
+    const outcome = await User.findByIdAndDelete(request.params.id);
     
     response.json({
-      User: result
+      User: outcome
     });
     
   } catch (err) {
     errorSwtich(err, response);
   }
 });
-
 
 
 module.exports = router;
